+++ pkg/deployment/resources/inspector/throttles.go
@@ -33,6 +33,7 @@ func NewDefaultThrottle() throttle.Components {
 		30*time.Second, // ArangoTask
 		30*time.Second, // ArangoRoute
 		30*time.Second, // ArangoProfile
+		30*time.Second, // ArangoPlatformStorage
 		30*time.Second, // Node
 		30*time.Second, // PV
 		15*time.Second, // PVC